requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
pyarabic==0.6.15
camel-tools==1.5.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0
ollama==0.3.3
ai21==2.6.0
fastapi==0.111.1
uvicorn[standard]==0.30.1
gradio==4.40.0
python-dotenv==1.0.1
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
transformers==4.42.4
httpx==0.27.0
pydantic==2.8.2
